+++ plugins/sudoers/log_client.h
@@ -51,9 +51,9 @@
 struct connection_buffer {
     TAILQ_ENTRY(connection_buffer) entries;
     uint8_t *data;
-    unsigned int size;
-    unsigned int len;
-    unsigned int off;
+    size_t size;
+    size_t len;
+    size_t off;
 };
 TAILQ_HEAD(connection_buffer_list, connection_buffer);
